third_party/WebKit/LayoutTests/external/wpt/css/cssom/CSSKeyframeRule-expected.txt

This is a testharness.js-based test.
PASS CSSKeyframeRule: style property
FAIL CSSKeyframeRule: style property has [PutForwards] assert_equals: margin-left expected "50%" but got "100%"
Harness: the test ran to completion.
